refactor(frontend): push state into decomposed components
