+++ frontend/README.md
@@ -17,7 +17,7 @@ This is the frontend admin application for the Lift Simulator system. It provide
 - Managing simulation runs through decomposed filter, table, pagination, and action-modal components, including multi-select bulk cancellation for active runs and bulk deletion for completed runs
 - Monitoring system health
 - Running simulations through decomposed setup, run-status, and results panels shared with run detail views
-- Maintaining scenarios and lift-system versions through decomposed page sections for easier UI maintenance
+- Maintaining scenarios and lift-system versions through decomposed page sections whose local UI state stays with the owning component or hook for easier maintenance
 
 For the overall project setup (backend, database, and Quick Start), see the [root README](../README.md). Backend REST API conventions (auth, RBAC, rate limits, error shape) live in [docs/API.md](../docs/API.md), and the always-current endpoint reference is the generated Swagger UI at `/api/v1/swagger-ui.html`.
