+++ pom.xml
@@ -14,7 +14,7 @@
 
     <groupId>com.liftsimulator</groupId>
     <artifactId>lift-simulator</artifactId>
-    <version>0.57.16</version>
+    <version>0.57.17</version>
     <packaging>jar</packaging>
 
     <name>Lift Simulator</name>
